# Speed-up array creation