--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -1,4 +1,4 @@
-pylint
+pylint<2.13.0
 pytest
 pytest-cov
 pytest-benchmark